REALLY fixed imports, reorganized slightly the code (utilities folder) for lisibility
