--- src/torchFastText/__init__.py
@@ -1,6 +1,0 @@
-import datasets
-import explainability
-import model
-import preprocess
-
-__all__ = ["datasets", "explainability", "model", "preprocess"]
