+++ src/torchFastText/model/lightning_module.py
@@ -1,7 +1,8 @@
 import pytorch_lightning as pl
 import torch
-from model.pytorch_model import FastTextModel
 from torchmetrics import Accuracy
+
+from .pytorch_model import FastTextModel
 
 
 class FastTextModule(pl.LightningModule):
