+++ src/torchFastText/model/pytorch_model.py
@@ -8,9 +8,10 @@ from typing import List
 import numpy as np
 import torch
 from captum.attr import LayerIntegratedGradients
-from preprocess import clean_text_feature
 from torch import nn
-from utils import (
+
+from ..preprocess import clean_text_feature
+from ..utilities.utils import (
     compute_preprocessed_word_score,
     compute_word_score,
     explain_continuous,
