+++ src/torchFastText/torchFastText.py
@@ -4,11 +4,6 @@ import time
 import numpy as np
 import pytorch_lightning as pl
 import torch
-from checkers import check_X, check_Y
-from datasets.dataset import FastTextModelDataset
-from datasets.tokenizer import NGramTokenizer
-from model.lightning_module import FastTextModule
-from model.pytorch_model import FastTextModel
 from pytorch_lightning.callbacks import (
     EarlyStopping,
     LearningRateMonitor,
@@ -16,6 +11,12 @@ from pytorch_lightning.callbacks import (
 )
 from pytorch_lightning.utilities import rank_zero_only
 from torch.optim import SGD, Adam
+
+from .datasets.dataset import FastTextModelDataset
+from .datasets.tokenizer import NGramTokenizer
+from .model.lightning_module import FastTextModule
+from .model.pytorch_model import FastTextModel
+from .utilities.checkers import check_X, check_Y
 
 logger = logging.getLogger(__name__)
