+++ src/torchFastText/utilities/__init__.py
@@ -1,0 +1,3 @@
+"""
+Init script.
+"""
