+++ src/torchFastText/utilities/checkers.py
